Core stages complete for eleven projects; flowlist through surface
~3,400 engine tests passing across the eleven finished cores. Each ships domain
models with DATA_MODEL invariants enforced, the pure engine implementing its
documented hard part, offline+live adapters, and SQLite plus in-memory
repositories. ethos core still building; surface and harden stages pending.

+++ projects/grailtrader/src/grailtrader/ids.py
@@ -40,7 +40,7 @@ def canonical_json(obj: Any) -> str:
 
 def cents(usd: float) -> int:
     """Serialise a USD amount as integer cents (no float repr enters a hash)."""
-    return int(round(usd * 100))
+    return round(usd * 100)
 
 
 def listing_id(source: str, external_id: str) -> str:
